+++ src/main/java/school/faang/bjs2_92935/Potion.java
@@ -1,10 +1,11 @@
 package school.faang.bjs2_92935;
 
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 
-
+@Getter
 @AllArgsConstructor
 public class Potion {
-    protected String name;
-    protected int requiredIngredients;
+    private final String name;
+    private final int requiredIngredients;
 }
